0.9.5

+++ script.steam.launcher/resources/scripts/steam-launch.sh
@@ -2,6 +2,30 @@
 #Script to launch Steam BPM from Xbmc, by teeedubb
 export DISPLAY=:0
 
+case "$(uname -s)" in
+    Darwin)
+        XBMC_PID=$(ps -A | grep XBMC.app | grep -v Helper | grep -v grep | awk '{print $1}')
+        XBMC_BIN=$(ps -A | grep XBMC.app | grep -v Helper | grep -v grep | awk '{print $5}')
+        STEAM_PID=$(ps -A | grep STEAM.app | grep -v Helper | grep -v grep | awk '{print $1}')
+        STEAM_BIN=$(ps -A | grep STEAM.app | grep -v Helper | grep -v grep | awk '{print $5}')
+        # Is Steam running?
+	if [[ $STEAM_PID ]]; then
+	  $STEAM_BIN steam://open/bigpicture #steam is brought to focus better this way
+	  echo "Steam already running"
+	else
+	   $STEAM_BIN -bigpicture #steam opens better like this if not already open
+	    echo "Steam not running, launching"
+	fi
+	# Wait for Steam to exit
+	while [ $(ps -A | grep STEAM.app | grep -v Helper | grep -v grep | awk '{print $1}') ]; do #STEAM_PID variable doesnt work here, needs work
+	  echo "Steam running"
+	  sleep 1
+	done
+	#Restart XBMC
+	$XBMC_BIN &
+        ;;
+    Linux)
+
 if [[ $(pidof steam) ]]; then
     if [[ $(wmctrl -l | grep "Steam$") ]]; then
       wmctrl -i -a $(wmctrl -l | grep "Steam$" | cut -c1-10) &
@@ -15,14 +39,22 @@ fi
 for i in {1..30} ; do
     if [[ $(wmctrl -l | grep "Steam$") ]]; then
 	if [[ $(pidof xbmc.bin) ]] ; then
-	  kill -9 $(pidof xbmc.bin)
+	  if [[ $3 = 0 ]]; then
+	    kill -9 $(pidof xbmc.bin)
+	  else
+	    wmctrl -r "XBMC Media Center" -b remove,fullscreen
+	  fi
 	fi
     fi
     sleep 0.1
 done
 
 if [[ $(pidof xbmc.bin) ]]; then
-	kill -9 $(pidof xbmc.bin)
+	  if [[ $3 = 0 ]]; then
+	    kill -9 $(pidof xbmc.bin)
+	  else
+	    wmctrl -r "XBMC Media Center" -b remove,fullscreen
+	  fi
 fi
 
 until [[ $(wmctrl -l | tail -n1 | grep "Steam$") ]]; do
@@ -40,6 +72,16 @@ done
 
 if [[ $(pidof xbmc.bin) ]] ; then
     wmctrl -a "XBMC Media Center"
+    if [[ $3 != 0 ]]; then
+      wmctrl -r "XBMC Media Center" -b add,fullscreen
+    fi
 else
     "$2" &
 fi
+
+        ;;    
+    *)
+        echo "I don't support this OS!"
+        exit 1
+        ;;
+esac
